Android Support and Money
Implemented basic android support, which was mostly in the forms of
getting the correct drivers, SDKs and all that jazz. Also working on
implementing the money system. Buying towers now costs money, as well as
enemys giving the player money for killing them. This only happens when
they are killed by a tower, not when they reach the end. This needs to
be changed to getters and setters though, so money is correctly updated.